refactor: organize tags dir
