+++ src/state/redux/profiles/selectors/tags/types.ts
@@ -1,0 +1,5 @@
+import { ITag } from '@core/features/tags';
+
+export type TagNode = Pick<ITag, 'id' | 'name'> & {
+	childrens?: TagNode[];
+};
